image_search_service/requirements.txt
# requirements.txt
fastapi
uvicorn
numpy
Pillow
requests
tensorflow
scikit-learn
pymongo
python-dotenv